flask-api/requirements.txt
flask>=2.3.3
flask-smorest==0.42.1
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-jwt-extended
passlib
flask-migrate
gunicorn
psycopg2
requests
# redis queue (celery can also be used)
redis
rq